backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# WebSocket support
websockets==12.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Redis for caching and sessions
redis==5.0.1
aioredis==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0
pydantic-settings==2.0.3

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
